tweak structure
