+++ challenge/challenge.rb
@@ -1,0 +1,2 @@
+require_relative './parser'
+require_relative './html_file_parser'
