+++ challenge/html_file_parser.rb
@@ -1,0 +1,14 @@
+require 'Nokogiri'
+
+module Challenge
+  class HtmlFileParser
+    def initialize(filepath)
+      @filepath = filepath
+    end
+
+    def parse
+      html = File.read(@filepath)
+      Nokogiri::HTML html
+    end
+  end
+end
